+++ now_lms/templates/learning/resources/slide_show.html
@@ -1,0 +1,33 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<title>reveal.js - Barebones</title>
+		<link rel="stylesheet" href="{{ url_for('static', filename='node_modules/reveal.js/dist/reveal.css') }}">
+		<link rel="stylesheet" href="{{ url_for('static', filename='node_modules/reveal.js/dist/theme/black.css') }}" id="theme">
+	</head>
+	<body>
+
+		<div class="reveal">
+			<div class="slides">
+
+				<section>
+					<h2>Barebones Presentation</h2>
+					<p>This example contains the bare minimum includes and markup required to run a reveal.js presentation.</p>
+				</section>
+
+				<section>
+					<h2>No Theme</h2>
+					<p>There's no theme included, so it will fall back on browser defaults.</p>
+				</section>
+
+			</div>
+		</div>
+
+		<script src="{{ url_for('static', filename='node_modules/reveal.js/dist/reveal.js') }}"></script>
+		<script>
+			Reveal.initialize();
+		</script>
+
+	</body>
+</html>
