+++ now_lms/templates/learning/resources/type_slides.html
@@ -38,32 +38,9 @@
 
           <div class="col-sm-8">
 
-            <div id="resource_container" class="d-flex justify-content-center">
-              
-              <iframe title="{{ recurso.nombre }}" id="Iframe" src="{{ url_for('external_code', recurso_code=recurso.codigo, course_code=recurso.curso) }}" frameborder="0"></iframe>
-
-              <script>
-                // Selecting the iframe element
-                var frame = document.getElementById("Iframe");
-                  
-                // Adjusting the iframe height onload event
-                frame.onload = function()
-                // function execute while load the iframe
-                {
-                  // set the height of the iframe as 
-                  // the height of the iframe content
-                  frame.style.height = 
-                  frame.contentWindow.document.body.scrollHeight + 'px';  
-
-                  // set the width of the iframe as the 
-                  // width of the iframe content
-                  frame.style.width  = 
-                  frame.contentWindow.document.body.scrollWidth+'px';
-                }
-                </script>
+            <iframe title="{{ recurso.nombre }}" id="Iframe" width="100%" height="315" src="{{ url_for('slide_show', recurso_code=recurso.codigo) }}" frameborder="0"></iframe>
 
           </div>
-        </div>
 
           <div class="col-sm-4">
