+++ now_lms/templates/learning/resources/type_youtube.html
@@ -37,7 +37,7 @@
           
           <div class="col-sm-8">
 
-            <iframe width="560" height="315" src="https://www.youtube.com/embed/{{ recurso.url[32:45]  }}" title="YouTube video player" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+            <iframe width="100%" height="315" src="https://www.youtube.com/embed/{{ recurso.url[32:45]  }}" title="YouTube video player" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 
           </div>
